Stabilize TestPgWrapper: correctly shut down backends on SIGTERM, and fix pointer corruption caused by implicit function declarations
Summary:
- Add a missing call to YBOnPostgresBackendShutdown to proc_exit, which fixes shutdown on SIGTERM.
- Tracked down an issue that manifested as YBCNewConstant returning a 64-bit pointer with high 4 bytes being overwritten with all zeros or all 0xff. This was because of a missing #include in ybc_fdw.c and the implicit function declaration assuming YBCNewConstant was returning an int. Trying to add extra flags like -Werror=implicit-function-declaration and -Werror=int-conversion where that won't break PostgreSQL code compilation.
- Fix a bug in BasePgSQLTest that resulted in creating ten PostgreSQL JDBC connections where one would be sufficient.
- Add TSAN suppressions for false positive signal-unsafe errors in PostgreSQL de-initialization.
- Add LSAN suppressions for what it thinks are memory leaks in PostgreSQL code.
- Add UBSAN suppressions for errors in random number generation / LZ compression code in PostgreSQL code. This also required updating the compiler flags.

Other improvements:
- Raise errors on ASAN/TSAN messages in mini-cluster logs.
- Post-process core files generated by postgres and initdb.
- Always destroy the YB mini-cluster after each test in BasePgSQLTest, even if PostgreSQL shutdown failed.
- Generate a combined compile_commands.json file for for both PostgreSQL and YB code, allowing to enable code completion in editors like Visual Studio Code.
- Enable Ninja by default when it is available, even if YB_USE_NINJA is not set.
- Moved some Java code out of TestUtils in the process of doing the above.
- Simplified the logic for counting retries when invoking the compiler on a remote host over ssh.

Test Plan: Jenkins: skip

Reviewers: neil, robert, mihnea

Reviewed By: mihnea

Subscribers: ybase

Differential Revision: https://phabricator.dev.yugabyte.com/D5502

+++ java/yb-client/src/test/java/org/yb/minicluster/MiniYBCluster.java
@@ -38,6 +38,7 @@ import org.yb.client.YBClient;
 import org.yb.util.NetUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.yb.util.RandomNumberUtil;
 
 import java.io.*;
 import java.net.InetAddress;
@@ -243,7 +244,7 @@ public class MiniYBCluster implements AutoCloseable {
     assert(TestUtils.IS_LINUX);
     // On Linux we can use 127.x.y.z, so let's just pick a random address.
     final StringBuilder randomLoopbackIp = new StringBuilder("127");
-    final Random rng = TestUtils.getRandomGenerator();
+    final Random rng = RandomNumberUtil.getRandomGenerator();
     for (int i = 0; i < 3; ++i) {
       // Do not use 0 or 255 for IP components.
       randomLoopbackIp.append("." + (1 + rng.nextInt(254)));
@@ -640,7 +641,7 @@ public class MiniYBCluster implements AutoCloseable {
     final MiniYBDaemon daemon =
         new MiniYBDaemon(type, indexForLog, command, proc, bindIp, rpcPort, webPort, cqlWebPort,
                          redisWebPort, pgsqlWebPort, dataDirPath);
-    logPrinters.addAll(daemon.getLogPrinters());
+    logPrinters.add(daemon.getLogPrinter());
 
     Thread.sleep(300);
     try {
@@ -660,7 +661,7 @@ public class MiniYBCluster implements AutoCloseable {
     String[] command = daemon.getCommandLine();
     LOG.info("Restarting process: {}", Joiner.on(" ").join(command));
     daemon = daemon.restart();
-    logPrinters.addAll(daemon.getLogPrinters());
+    logPrinters.add(daemon.getLogPrinter());
 
     Process proc = daemon.getProcess();
     Thread.sleep(300);
@@ -701,25 +702,55 @@ public class MiniYBCluster implements AutoCloseable {
     LOG.info("Restarted mini cluster");
   }
 
-  private void processCoreFile(MiniYBDaemon daemon) throws Exception {
-    final int pid = daemon.getPid();
-    final File coreFile = new File("core." + pid);
-    if (coreFile.exists()) {
-      LOG.warn("Found core file '{}' from {}", coreFile.getAbsolutePath(), daemon.toString());
-      Process analyzeCoreFileScript = new ProcessBuilder().command(Arrays.asList(new String[]{
-          TestUtils.findYbRootDir() + "/build-support/analyze_core_file.sh",
-          "--core",
-          coreFile.getAbsolutePath(),
-          "--executable",
-          daemon.getCommandLine()[0]
-      })).inheritIO().start();
-      analyzeCoreFileScript.waitFor();
-      if (coreFile.delete()) {
-        LOG.warn("Deleted core file at '{}'", coreFile.getAbsolutePath());
-      } else {
-        LOG.warn("Failed to delete core file at '{}'", coreFile.getAbsolutePath());
+  public static void processCoreFile(int pid,
+                                     String executablePath,
+                                     String programDescription,
+                                     File coreFileDir,
+                                     boolean tryWithoutPid) throws Exception {
+    List<String> coreFileNames = new ArrayList<String>();
+    coreFileNames.add("core." + pid);
+    if (tryWithoutPid) {
+      coreFileNames.add("core");
+    }
+    for (String coreBasename : coreFileNames) {
+      final File coreFile =
+          coreFileDir == null ? new File(coreBasename) : new File(coreFileDir, coreBasename);
+      if (coreFile.exists()) {
+        LOG.warn("Found core file '{}' from {}", coreFile.getAbsolutePath(), programDescription);
+        String analysisScript = TestUtils.findYbRootDir() + "/build-support/analyze_core_file.sh";
+        List<String> analysisArgs = Arrays.asList(
+            analysisScript,
+            "--core",
+            coreFile.getAbsolutePath(),
+            "--executable",
+            executablePath
+        );
+        LOG.warn("Analyzing core file using the command: " + analysisArgs);
+        ProcessBuilder procBuilder = new ProcessBuilder().command(analysisArgs);
+        procBuilder.redirectErrorStream(true);
+        Process analysisProcess = procBuilder.start();
+
+        LogPrinter logPrinter = new LogPrinter(analysisProcess.getInputStream(), "    ");
+        analysisProcess.waitFor();
+        logPrinter.stop();
+
+        if (analysisProcess.exitValue() != 0) {
+          LOG.warn("Core file analysis script " + analysisProcess + " exited with code: " +
+              analysisProcess.exitValue());
+        } else {
+          if (coreFile.delete()) {
+            LOG.warn("Deleted core file at '{}'", coreFile.getAbsolutePath());
+          } else {
+            LOG.warn("Failed to delete core file at '{}'", coreFile.getAbsolutePath());
+          }
+        }
       }
     }
+  }
+
+  private void processCoreFile(MiniYBDaemon daemon) throws Exception {
+    processCoreFile(daemon.getPid(), daemon.getCommandLine()[0], daemon.toString(),
+        /* coreFileDir */ null, /* tryWithoutPid */ false);
   }
 
   private void destroyDaemon(MiniYBDaemon daemon) throws Exception {
@@ -850,6 +881,7 @@ public class MiniYBCluster implements AutoCloseable {
    */
   public String getMasterAddresses() {
     return masterAddresses;
+
   }
 
   /**
